Flex001_JJ/requirements.txt

PySide6>=6.5.0
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
opencv-python>=4.8.0
pypylon>=2.0.0
pyserial>=3.5
numpy>=1.24.0
Pillow>=10.0.0



